// src/types/arinv.ts
export interface ARINV {
  ARINV_GUID: string
  RECNO5: string
  ID: string
  NAME: string
  ADDRESS1: string
  ADDRESS2: string | null
  CITY: string
  STATE: string
  ZIP: string
  PHONE: string
  INVOICE: string
  DESCR: string
  INV_DATE: string
  DUE_DATE: string
  DIS_DATE: string
  IDDISCOUNT: string
  IDCHARGE: string
  DISCOUN: string
  CHARGE: string
  TERMS: string
  IN_LEVEL: string
  PO_NO: string
  SALESMAN: string
  SHIP_VIA: string
  E_DATE: string
  DISCOUNT: string
  OVERDUE: string
  FREIGHT: string
  HANDLING: string
  TAX: string
  SUBTOTAL: string
  TOTAL: string
  STATUS: string
  TOTAL_PAID: string
  TOTAL_W: string
  BALANCE: string
  CHECK_NO: string
  CHECK_DATE: string
  C_ID: string
  C_NAME: string
  C_ADDRESS1: string
  C_ADDRESS2: string
  C_CITY: string
  C_STATE: string
  C_ZIP: string
  C_EMAIL: string
  C_PHONE: string
  EMAIL: string
  BILLING_CO: string
  SHIP_DATE: string
  FRGHT_AMT: string
  FRGHT_TYPE: string
  USER_NAME: string
  WAS_PROC: string
  MEMO: string
  INTERNALNT: string
  SIGNATURE: string
  COUNTRY: string
  C_COUNTRY: string
  CLOSE_DATE: string
  SO_TAX: string
  MAN_TAX: string
  ORIG_INV: string
  ORIG_AID: string
  PRINT_TIME: string
  C_FR_ORDER: string
  CREA_DATE: string
  CREA_TIME: string
  PROC_DATE: string
  PROC_TIME: string
  DIV_AID: string
  AR_TRADE: string
  FREIGHT_GL: string
  FINANCE: string
  CUST_DISC: string
  DOWN_PAY: string
  DEF_DEPT: string
  WAREHOUSE: string
  LAST_PRINT: string
  FREIGHTEXP: string
  LY_PTS_GOT: string
  OUT_CITY: string
  AUTOID: string
  CSEND_DATE: string
  CSEND_TIME: string
  CSEND_MES: string
  TO_B_SENT: string
  JOB_ID: string
  NUM_BILL: string
  USE_JOBRET: string
  FROM_TANDM: string
  SUBILNORET: string
  ARQTAIDSRC: string
  USETAXED: string
  CNTCT_AID: string
  CCNTCT_AID: string
  MAN_USETAX: string
  TRIM: string
  METAL: string
  CF_EMAIL: string
  CF_SENT: string
  ARTJ_EX_ID: string
  ROLLEXDT: string
  ROLLPROCDT: string
  EXTERNALID: string
  P_ROUNDDIF: string
  AUTHORIZED: string
  ARCUST_GUID: string
  TYPE: string
  TREE_ID: string
  L_NAME: string
  F_NAME: string
  SHIP_TO: string
  CONID_1: string
  CONTACT_1: string
  CONID_2: string
  CONTACT_2: string
  CONID_3: string
  CONTACT_3: string
  CONID_4: string
  CONTACT_4: string
  CONID_5: string
  CONTACT_5: string
  CRED_LIM: string
  IDGRACE: string
  GRACE: string
  FC_RATE: string
  MIN_FC: string
  DEF_FOB: string
  MAIL_LIST: string
  DIRECTIONS: string
  DEL_REASON: string
  EMAIL_LIST: string
  CASH: string
  CHECK: string
  CREDITCARD: string
  CHARGEPYMT: string
  CHARGEDFLT: string
  BANK_ROUTE: string
  SAVINGS: string
  BANK_DESCR: string
  BANK_ACCOU: string
  GROUPACHTR: string
  INACTIVE: string
  TREE_PATH: string
  GPS_ADD1: string
  GPS_ADD2: string
  GPS_CITY: string
  GPS_STATE: string
  GPS_ZIP: string
  STM_D: string
  STM_TIME: string
  STM_MES: string
  CRM_EXCLUD: string
  ARTJ_CHGD: string
  ROLLEXPDT: string
  LY_RATIO: string
  LY_CARD: string
  LY_DESCR: string
  LY_PPCLIID: string
  LY_SOFTWR: string
  LY_D_TIME: string
  LY_USER: string
  DEF_CLI_ID: string
  CUST_NOTE: string
  TREE_DESCR: string
}